Refactor scroller
